feat: updated code for project pages
