+++ src/components/molecules/WorkAccordionHeader/index.tsx
@@ -1,11 +1,7 @@
 import * as Chakra from "@chakra-ui/react";
 import { CustomText } from "../../atoms/CustomText";
-
-import { motion } from "framer-motion";
-import { HiOutlineChevronDown } from "react-icons/hi";
 import { BiMapPin } from "react-icons/bi";
-
-const MotionIcon = motion.create(Chakra.Icon);
+import { WorkAccordionChevron } from "../WorkAccordionChevron";
 
 interface IProps {
   title: string;
@@ -15,6 +11,7 @@ interface IProps {
   isOpen: boolean;
   handleChangeIsOpen: (bool: boolean) => void;
 }
+
 export const WorkAccordionHeader = ({
   title,
   company,
@@ -80,25 +77,10 @@ export const WorkAccordionHeader = ({
           </Chakra.Badge>
         </Chakra.HStack>
 
-        <Chakra.IconButton
-          position={{ base: "absolute", md: "static" }}
-          aria-label="Toggle Accordion"
-          onClick={() => handleChangeIsOpen(!isOpen)}
-          size="sm"
-          variant="ghost"
-          alignSelf={{ base: "flex-end", md: "center" }}
-          _hover={{ bg: "none" }}
-          _focus={{ boxShadow: "none", outline: "none" }}
-        >
-          <MotionIcon
-            as={HiOutlineChevronDown}
-            position={{ base: "absolute", md: "static" }}
-            color={"#FFF"}
-            animate={{ rotate: isOpen ? 180 : 0 }}
-            transition={{ duration: 0.3 }}
-            boxSize={6}
-          />
-        </Chakra.IconButton>
+        <WorkAccordionChevron
+          isOpen={isOpen}
+          handleChangeIsOpen={() => handleChangeIsOpen(!isOpen)}
+        />
       </Chakra.Flex>
     </Chakra.Box>
   );
