+++ src/components/organisms/ProductTechStack/index.tsx
@@ -1,3 +1,5 @@
+import { TechTag } from "@/components/atoms/TechTag";
+import { validateOptionsBasedOnBoolean } from "@/utils/validateOptionsBasedOnBoolean";
 import * as Chakra from "@chakra-ui/react";
 import { motion } from "framer-motion";
 
@@ -23,29 +25,14 @@ export const ProductTechStack = ({ technologies }: Props) => {
       >
         {technologies.map((tech, index) => (
           <Chakra.WrapItem key={tech}>
-            <Chakra.Box
-              bgGradient={
-                index % 2 === 0
-                  ? "linear-gradient(to bottom, #667eea, #764ba2)"
-                  : "linear-gradient(to bottom, #4facfe, #00f2fe)"
-              }
-              color="white"
-              px={5}
-              py={3}
-              borderRadius="full"
-              fontWeight="semibold"
-              fontSize="sm"
-              overflow="hidden"
-              boxShadow={"0 8px 25px rgba(102, 126, 234, 0.3)"}
-              transition="all 0.3s ease"
-              willChange="transform, box-shadow"
-              _hover={{
-                transform: "translateY(-3px)",
-                boxShadow: "0 12px 35px rgba(102, 126, 234, 0.4)",
-              }}
-            >
-              <Chakra.Text>{tech}</Chakra.Text>
-            </Chakra.Box>
+            <TechTag
+              bgGradient={validateOptionsBasedOnBoolean(
+                index % 2 === 0,
+                "linear-gradient(to bottom, #667eea, #764ba2)",
+                "linear-gradient(to bottom, #4facfe, #00f2fe)"
+              )}
+              tech={tech}
+            />
           </Chakra.WrapItem>
         ))}
       </Chakra.Wrap>
