+++ src/data/colors/index.ts
@@ -1,0 +1,42 @@
+export const SKILL_CATEGORY_COLORS = {
+  frontend: {
+    bg: "#E3F2FD",
+    border: "#90CAF9",
+    tag: {
+      bg: "#BBDEFB",
+      text: "#1565C0",
+      hoverBg: "#1E88E5",
+      hoverText: "#FFFFFF",
+    },
+  },
+  backend: {
+    bg: "#F3E8FD",
+    border: "#CEBDF0",
+    tag: {
+      bg: "#E0D3F5",
+      text: "#5E35B1",
+      hoverBg: "#7E57C2",
+      hoverText: "#FFFFFF",
+    },
+  },
+  testing: {
+    bg: "#FFF3E0",
+    border: "#FFCC80",
+    tag: {
+      bg: "#FFE0B2",
+      text: "#EF6C00",
+      hoverBg: "#FFA726",
+      hoverText: "#FFFFFF",
+    },
+  },
+  tools: {
+    bg: "#FFF0EB",
+    border: "#FFB6A9",
+    tag: {
+      bg: "#FFD6CC",
+      text: "#D84315",
+      hoverBg: "#FF7043",
+      hoverText: "#FFFFFF",
+    },
+  },
+};
